Wiped Old database info
I removed the two added entries from the database and checked that registration worked for the added fields.

I kept the master information so the admin panel is still functional.

I also changed the second address field to be optional along with NFP items since not all will have this.

-Catherine