NFP Registration
Started working on NFP registration. Set up in models, views, forms, and urls. Also started on the html for it.

Successfully redirected the initial registration of an NFP to a secondary registration page.

Cleared Database entries besides MASTER.

Current issue is in html, form fields aren't appearing on the page.